+++ src/main/kotlin/joshxviii/plantz/PazCriteria.kt
@@ -1,6 +1,7 @@
 package joshxviii.plantz
 
 import joshxviii.plantz.advancement.*
+import net.minecraft.advancements.CriterionProgress
 import net.minecraft.core.Registry
 import net.minecraft.core.registries.BuiltInRegistries