+++ src/main/kotlin/joshxviii/plantz/item/SunBatteryItem.kt
@@ -2,11 +2,13 @@ package joshxviii.plantz.item
 
 import joshxviii.plantz.PazBlocks
 import joshxviii.plantz.PazComponents
+import joshxviii.plantz.PazCriteria
 import joshxviii.plantz.PazItems
 import joshxviii.plantz.block.entity.MailboxBlockEntity
 import joshxviii.plantz.block.entity.SunBatteryBlockEntity
 import joshxviii.plantz.item.component.StoredSun
 import net.minecraft.core.BlockPos
+import net.minecraft.server.level.ServerPlayer
 import net.minecraft.util.Mth
 import net.minecraft.world.InteractionResult
 import net.minecraft.world.entity.LivingEntity